requirement.txt
adium-theme-ubuntu==0.3.4
chardet==2.3.0
cliapp==1.20160109
Django==1.11.4
django-autoslug==1.9.3
django-bootstrap3==9.0.0
django-braces==1.11.0
django-debug-toolbar==1.8
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
Markdown==2.6.6
olefile==0.44
Pillow==4.2.1
pycurl==7.43.0
Pygments==2.1
PyJWT==1.5.3
python-apt==1.1.0b1
pytz==2017.2
PyYAML==3.11
scour==0.32
six==1.10.0
sqlparse==0.2.3
ttystatus==0.32
unity-lens-photos==1.0
virtualenv==15.1.0
whitenoise==3.3.0
xlrd==1.0.0
